requirements.txt
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.1
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fsspec==2024.5.0
gpg==1.10.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging @ file:///croot/packaging_1710807400464/work
parso==0.8.4
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
Pygments @ file:///croot/pygments_1684279966437/work
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
pyzmq==26.0.3
shiboken2==5.15.2.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.6.3
sympy==1.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.12.0
wcwidth==0.2.13
